Retrieve data and buttons function
1. I fixed the data retrieve from local storage
issue by myLibrary = checks for storage.
if local storage has item myLibrary then I
populate the storage and return the existing library.

To populate storage, I retrieve data from localStorage
then parse it. For each item in myLibrary (array), I create
a tempBook using the tempBook function.
Create temp book will populate the divs onto the page

2. I added functionality for buttons remove read and unread
For read or Unread btns I created QSAll and for each
button I add an EL. If the EL hear's a click, it removes
the current "read or unread" value and adds "unread or read"
value. I also change myLibrary[i].read = "unread or read"
* I realised much later that myLibrary[i] is EQUAL to
readOrUnreadBtn[i] as they are from same position.(It does not
automatically refresh so, we need to location.reload)

Then clear storage and add the updated version of the library in.

For remove btn, it was a bit trickier bcoz I forgot my array stuff
I found that to use slice, you create 2 new arrays containing stuff
before the index you want to remove and after the index you want to
remove. create a final array to combine the 2 using ...frontPart,
...endPart. then assign that final array as myLibrary.
Clear the storage and update the storage's myLibrary. (It does not
automatically refresh so, we need to location.reload)

3. I fixed the footer from always floating up when I opened the modal.
I changed it to pos absolute to fix it completely.